backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
kagglehub
python-multipart
